+++ engine/src/main/java/dev/gamekit/ui/events/MouseMotionEvent.java
@@ -1,0 +1,11 @@
+package dev.gamekit.ui.events;
+
+/**
+ * A {@link MouseEvent} which is dispatched when
+ * the mouse is moved over a listener widget
+ */
+public class MouseMotionEvent extends MouseEvent {
+  public MouseMotionEvent(int x, int y) {
+    super(x, y);
+  }
+}